Migrate analyticsService to TypeScript
